Fix critical submission issues
- Update @raycast/api from 1.54.0 to 1.100.3 for compatibility
- Fix command names to follow Raycast guidelines (verb-noun structure):
  - 'Looma.fm Music' → 'Play Music'
  - 'Pause Looma.fm' → 'Pause Music'
  - 'Reshuffle Looma.fm' → 'Shuffle Music'
  - 'Stop Looma.fm' → 'Stop Music'
  - 'Short Meditation' → 'Start Meditation'
- Replace complex React component with simple function for reliability
- Remove unused backup files and clean imports
